ref: Hoist `flush`, `close`, and `lastEventId` into `@sentry/core` (#8731)
